+++ environment/vm/lima/limautil/limautil.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
 
 	"github.com/abiosoft/colima/cli"
@@ -89,45 +88,6 @@ func (i InstanceInfo) Running() bool { return i.Status == limaStatusRunning }
 // Config returns the current Colima config
 func (i InstanceInfo) Config() (config.Config, error) {
 	return configmanager.LoadFrom(ColimaStateFile(i.Name))
-}
-
-// ShowSSH runs the show-ssh command in Lima.
-// returns the ssh output, if in layer, and an error if any
-func ShowSSH(profileID string) (resp struct {
-	Output string
-	File   struct {
-		Lima   string
-		Colima string
-	}
-}, err error) {
-	ssh := sshConfig(profileID)
-	sshConf, err := ssh.Contents()
-	if err != nil {
-		return resp, fmt.Errorf("error retrieving ssh config: %w", err)
-	}
-
-	resp.Output = replaceSSHConfig(sshConf, profileID)
-	resp.File.Lima = ssh.File()
-	resp.File.Colima = config.SSHConfigFile()
-	return resp, nil
-}
-
-func replaceSSHConfig(conf string, profileID string) string {
-	profileID = config.Profile(profileID).ID
-
-	var out bytes.Buffer
-	scanner := bufio.NewScanner(strings.NewReader(conf))
-
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		if strings.HasPrefix(line, "Host ") {
-			line = "Host " + profileID
-		}
-
-		_, _ = fmt.Fprintln(&out, line)
-	}
-	return out.String()
 }
 
 // Lima statuses
@@ -244,39 +204,4 @@ func LimaHome() string {
 	}
 
 	return config.LimaDir()
-}
-
-const colimaStateFileName = "colima.yaml"
-
-// ColimaStateFile returns path to the colima state yaml file.
-func ColimaStateFile(profileID string) string {
-	return filepath.Join(LimaHome(), config.Profile(profileID).ID, colimaStateFileName)
-}
-
-const colimaDiffDisk = "diffdisk"
-
-// ColimaDiffDisk returns path to the diffdisk for the colima VM.
-func ColimaDiffDisk(profileID string) string {
-	return filepath.Join(LimaHome(), config.Profile(profileID).ID, colimaDiffDisk)
-}
-
-const sshConfigFile = "ssh.config"
-
-// sshConfig is the ssh configuration file for a Colima profile.
-type sshConfig string
-
-// Contents returns the content of the SSH config file.
-func (s sshConfig) Contents() (string, error) {
-	profile := config.Profile(string(s))
-	b, err := os.ReadFile(s.File())
-	if err != nil {
-		return "", fmt.Errorf("error retrieving Lima SSH config file for profile '%s': %w", strings.TrimPrefix(profile.DisplayName, "lima"), err)
-	}
-	return string(b), nil
-}
-
-// File returns the path to the SSH config file.
-func (s sshConfig) File() string {
-	profile := config.Profile(string(s))
-	return filepath.Join(LimaHome(), profile.ID, sshConfigFile)
 }
